Create backend and frontend for user auth
Utilized React/Redux to handle frontend auth. Users are dropped at the home page. Users have two options, sign in or sign up. From sign up users can create accounts, and are redirected to the homepage. Once signed in, users are presented with a logout button. From sign in, users can create a session and are redirected back to the home page. Users are presented with a logout button.

Undesired consequence of the current frontend routes:

 - Headers and footer are rendering on login and signup pages