Add params back to permitted product params
Previously removed :gender, :country, and :last_name from product params
as they are not used in the app, however they are needed to create new
users.

These params are not currently used in the app. In addition, gender and
country are set to defaults. Instead of removing the columns, logic, and
params decided to leave as is. There are plans to use gender to auto
filter product index and last name and country can be leveraged in the
cart component.