Add Order Item Count to Bag Icon on NavToolbar
User will now have indication add to cart was successful, effects are
scrolling to top & update count on the shopping icon of NavToolBar.

Now fetching Orders on NavBarContainer, which will allow a updated
count on any page. Additionally, no longer updating products state
slice to single product for /products/:id as this caused and error
when navigating to cart from that location.